+++ pysat/utils/misc.py
@@ -51,43 +51,6 @@ def computational_form(data):
     return dslice
 
 
-def set_data_dir(path=None, store=None):
-    """
-    Set the top level directory pysat uses to look for data and reload.
-
-    Parameters
-    ----------
-    path : string
-        valid path to directory pysat uses to look for data
-    store : bool
-        if True, store data directory for future runs
-    """
-    import sys
-    import os
-    import pysat
-    if sys.version_info[0] >= 3:
-        if sys.version_info[1] < 4:
-            import imp
-            re_load = imp.reload
-        else:
-            import importlib
-            re_load = importlib.reload
-    else:
-        re_load = reload
-    if store is None:
-        store = True
-    if os.path.isdir(path):
-        if store:
-            with open(os.path.join(os.path.expanduser('~'), '.pysat',
-                                   'data_path.txt'), 'w') as f:
-                f.write(path)
-        pysat.data_dir = path
-        pysat._files = re_load(pysat._files)
-        pysat._instrument = re_load(pysat._instrument)
-    else:
-        raise ValueError('Path %s does not lead to a valid directory.' % path)
-
-
 def load_netcdf4(fnames=None, strict_meta=False, file_format=None,
                  epoch_name='Epoch', units_label='units',
                  name_label='long_name', notes_label='notes',
@@ -412,78 +375,3 @@ def load_netcdf4(fnames=None, strict_meta=False, file_format=None,
         out.append(pds.DataFrame.from_records(item, index=epoch_name))
     out = pds.concat(out, axis=0)
     return out, mdata
-
-
-def scale_units(out_unit, in_unit):
-    """ Determine the scaling factor between two units
-
-    Parameters
-    -------------
-    out_unit : str
-        Desired unit after scaling
-    in_unit : str
-        Unit to be scaled
-
-    Returns
-    -----------
-    unit_scale : float
-        Scaling factor that will convert from in_units to out_units
-
-    Notes
-    -------
-    Accepted units include degrees ('deg', 'degree', 'degrees'),
-    radians ('rad', 'radian', 'radians'),
-    hours ('h', 'hr', 'hrs', 'hour', 'hours'), and lengths ('m', 'km', 'cm').
-    Can convert between degrees, radians, and hours or different lengths.
-
-    Example
-    -----------
-    ::
-    import numpy as np
-    two_pi = 2.0 * np.pi
-    scale = scale_units("deg", "RAD")
-    two_pi *= scale
-    two_pi # will show 360.0
-
-
-    """
-
-    if out_unit == in_unit:
-        return 1.0
-
-    accepted_units = {'deg': ['deg', 'degree', 'degrees'],
-                      'rad': ['rad', 'radian', 'radians'],
-                      'h': ['h', 'hr', 'hrs', 'hours'],
-                      'm': ['m', 'km', 'cm'],
-                      'm/s': ['m/s', 'cm/s', 'km/s']}
-
-    scales = {'deg': 180.0, 'rad': np.pi, 'h': 12.0,
-              'm': 1.0, 'km': 0.001, 'cm': 100.0,
-              'm/s': 1.0, 'cm/s': 100.0, 'km/s': 0.001}
-
-    # Test input and determine transformation type
-    out_key = None
-    in_key = None
-    for kk in accepted_units.keys():
-        if out_unit.lower() in accepted_units[kk]:
-            out_key = kk
-        if in_unit.lower() in accepted_units[kk]:
-            in_key = kk
-
-    if out_key is None:
-        raise ValueError('Unknown output unit {:}'.format(out_unit))
-
-    if in_key is None:
-        raise ValueError('Unknown input unit {:}'.format(in_unit))
-
-    if out_key == 'm' or out_key == 'm/s' or in_key == 'm' or in_key == 'm/s':
-        if in_key != out_key:
-            raise ValueError('Cannot scale {:s} and {:s}'.format(out_unit,
-                                                                 in_unit))
-        # Recast units as keys for the scales dictionary
-        out_key = out_unit
-        in_key = in_unit
-
-    unit_scale = scales[out_key.lower()] / scales[in_key.lower()]
-
-    return unit_scale
